[tg-45] Implementa backoffice institución y sector.

+++ back-end/hospital-api/src/main/java/net/pladema/person/controller/dto/APersonDto.java
@@ -67,7 +67,7 @@ public class APersonDto {
 
     private String quarter;
 
-    private Short cityId;
+    private Integer cityId;
 
     private String postcode;
